feat: update default theme and rename it to default component
