+++ flottform/forms/src/index.ts
@@ -4,4 +4,4 @@ export { FlottformTextInputClient } from './flottform-text-input-client';
 export { FlottformTextInputHost } from './flottform-text-input-host';
 export type { ClientState, SafeEndpointInfo, Logger } from './internal';
 import './theme/style.css';
-export { defaultThemeForFileInput } from './theme/defaultTheme';
+export { createDefaultFlottformComponent } from './theme/flottform-default-component';
